pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["jraphaelsst <joaoraphaelsst@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.4.3"
factory-boy = "^3.3.0"
django = "^5.0"
django-rest-framework = "^0.1.0"
pytest-django = "^4.7.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.2.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bookstore.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
